feat: reconstruct distributed

+++ internal/server/api_varz.go
@@ -115,8 +115,8 @@ type Varz struct {
 	Mem                 int64   `json:"mem"`                   // 内存
 	CPU                 float64 `json:"cpu"`                   // cpu
 	Goroutine           int     `json:"goroutine"`             // goroutine数量
-	ClusterReqGoroutine int     `json:"cluster_req_goroutine"` // 分布式请求goroutine数量
-	ClusterMsgGoroutine int     `json:"cluster_msg_goroutine"` // 分布式消息goroutine数量
+	ClusterReqGoroutine int64   `json:"cluster_req_goroutine"` // 分布式请求goroutine数量
+	ClusterMsgGoroutine int64   `json:"cluster_msg_goroutine"` // 分布式消息goroutine数量
 
 	InMsgs      int64 `json:"in_msgs"`      // 流入消息数量
 	OutMsgs     int64 `json:"out_msgs"`     // 流出消息数量
